+++ Kiki.Courier/Kiki.Courier.Domain/Trip.cs
@@ -8,8 +8,8 @@ namespace Kiki.Courier.Domain
 {
     public class Trip
     {
-        public decimal StartTime { get; set; }
-        public decimal EndTime { get; set; }
+        public double StartTime { get; set; }
+        public double EndTime { get; set; }
         public int Distance { get; set; }
         public List<Package> Packages { get; set; } = new List<Package>();
     }
